+++ CHANGELOG.md
@@ -38,13 +38,13 @@
 - Removed deprecated `limit` on `CreateEventSubSubscription`
 - Removed deprecated `AddBlockedTerm`, `DeleteBlockedTerm`, `DeletePermittedTerm` and `AddPermittedTerm` on `ModerationActionCommand`
 - Deprecated TMI
-- Added `tags` to Modify Channel Information, Get Channel Information, Search Channels, Get Followed Streams and Get Streams
 - Deprecated `channel.follow` v1 eventsub event
 - Deprecated `Get User Follows` and associated follower related extension methods
 - Deprecated Twitch-defined tags: `Get All Stream Tags`, `Get Stream Tags`, `Replace Stream Tags` and `TwitchTag`
 
 ### Added
 
+- Added `tags` to `Modify Channel Information`, `Get Channel Information`, `Search Channels`, `Get Followed Streams` and `Get Streams`
 - Added `beta` feature to specifically enable beta endpoints
 - Added `Ban User` and `Unban User`
 - Added `Get Chat Settings` endpoint
@@ -69,6 +69,7 @@
 - Added `channel.shield_mode.begin` and `channel.shield_mode.end` EventSub events
 - Added `tower` implementation for `HttpClient`
 - Added `helix::Response::new` and `helix::Response::with_data` to make it possible to create your own responses.
+- Added `is_branded_content` and `content_classification_labels` to `Get Channel Information` and `Modify Channel information`
 
 ### Fixed